# backend/openapi.yaml
openapi: 3.0.3
info:
  title: MyPetCare API
  version: "1.0.0"
  description: |
    Backend REST API per MyPetCare - Piattaforma di prenotazione servizi per animali domestici
    
    ## Features
    - ✅ Authentication via Firebase Auth
    - ✅ Booking management (create, cancel, history)
    - ✅ PRO directory with geospatial search
    - ✅ Reviews and ratings
    - ✅ Payment processing (Stripe/PayPal)
    - ✅ GDPR compliance (data export & deletion)
    - ✅ Real-time messaging (chat support)
    
    ## Rate Limiting
    - General API: 100 requests/15min
    - Authentication: 5 requests/15min
    - GDPR endpoints: 10 requests/hour
    
  contact:
    name: MyPetCare API Support
    email: support@mypetcareapp.org
  license:
    name: Proprietary
servers:
  - url: https://api.mypetcareapp.org
    description: Production server
  - url: https://api-staging.mypetcareapp.org
    description: Staging server
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase Authentication ID token
  schemas:
    Error:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        message:
          type: string
          example: "Resource not found"
        code:
          type: string
          example: "NOT_FOUND"
...
    Booking:
      type: object
      required: [id, proId, userId, date, status]
      properties:
        id:
          type: string
          example: "booking_789012"
        proId:
          type: string
          example: "pro_123456"
        userId:
          type: string
          example: "user_345678"
        date:
          type: string
          format: date-time
          example: "2024-12-15T10:00:00Z"
        serviceId:
          type: string
          example: "service_grooming_basic"
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
          example: "confirmed"
        notes:
          type: string
          example: "Il mio cane ha paura dei rumori forti"
        totalAmount:
          type: number
          format: double
          example: 45.00
        currency:
          type: string
          example: "EUR"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /api/bookings:
    post:
      summary: Crea prenotazione
      description: Crea una nuova prenotazione per un servizio PRO
      tags: [Bookings]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [proId, date, serviceId]
              properties:
                proId:
                  type: string
                  example: "pro_123456"
                date:
                  type: string
                  format: date-time
                  example: "2024-12-15T10:00:00Z"
                serviceId:
                  type: string
                  example: "service_grooming_basic"
                notes:
                  type: string
                  example: "Il mio cane ha paura dei rumori forti"
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Booking'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/bookings/{bookingId}:
    get:
      summary: Dettagli prenotazione
      tags: [Bookings]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: bookingId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Booking details
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Booking'
        '404':
          description: Booking not found

    delete:
      summary: Cancella prenotazione
      tags: [Bookings]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: bookingId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Booking cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Booking cancelled successfully"
...
security:
  - BearerAuth: []